Moved out collapse revissions as well
